+++ functions/src/index.ts
@@ -2,6 +2,7 @@ import { handlerHelper } from "./helpers/handlerHelper.js";
 import { moonpaySignFunction } from "./functions/moonpay.sign.js";
 import { banxaAuthFunction } from "./functions/banxa.auth.js";
 import { sentryTunnelFunction } from "./functions/sentry.tunnel.js";
+import { mixpanelTunnelFunction } from "./functions/mixpanel.tunnel.js";
 
 /**
  * Banxa Authentication server-side lambda
@@ -14,6 +15,12 @@ export const banxaAuth = handlerHelper(banxaAuthFunction);
  * @function moonpay_sing
  */
 export const moonpaySign = handlerHelper(moonpaySignFunction);
+
+/**
+ * Proxy requests to mixpanel with lambda
+ * @function mixpanelProxy
+ */
+export const mixpanelTunnel = handlerHelper(mixpanelTunnelFunction);
 
 
 /**